+++ DOCS/CHANGELOG.md
@@ -5,10 +5,10 @@
 ### Added
 - new features: `_maxest`, `_value_all`, `_value[8|16|32|64|128|256|512|1024]`.
 - new traits `DataValue[Copy]`, `DataType[Copy]`, `DataRaw[Copy]`, `DirEnv`, `ExtLog`, `ExtProcess`.
-- new types: `DataValue*`, `DataType*`, `DataRaw*`, `DirApple`, `DirWindows`, `DirUnix`, `DirXdg`, `LogConfig`, `NoData`, .
+- new types: `DataValue*`, `DataType*`, `DataRaw*`, `DirApple`, `DirWindows`, `DirUnix`, `DirXdg`, `LogConfig`, `NoData`, `Base`, `Crockford`, `Rfc4648`, `Rfc4648Hex`.
   - namespaces: `Iter`, `Log`.
 - new macros: `maybe!`, `xorshift_custom!`.
-- new modules: `data::{codec, list, key, table, uid}`, `lang::{c}`, `sys::log`.
+- new modules: `data::{codec::{self, radix}, list, key, table, uid}`, `lang::{c}`, `sys::log`.
 - new hidden const arrays: `XOROSHIFT_[16|32|64]_TRIPLETS`.
 - new macro arms:
   - `str!`: `ip_addr`.